% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_garch.R
\name{get_u_from_garch}
\alias{get_u_from_garch}
\alias{get_q_from_garch}
\title{PIT for GARCH models and its inverse.}
\usage{
get_u_from_garch(models)

get_q_from_garch(u, models, time_index)
}
\arguments{
\item{models}{a list of models returned from [fGarch::garchFit()].}

\item{u}{a matrix of evaluation points.}

\item{time_index}{the index for the time point in the original data.}
}
\value{
A matrix of conditional quantiles from the garch models (conditioned
    on `time == time_s[time_index]`).
}
\description{
Applies the conditional marginal distribution (from GARCH) to the residuals.
}
\examples{
fit <- select_garch(rnorm(100))
get_u_from_garch(fit)

}
