% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_garch.R
\name{select_garch}
\alias{select_garch}
\title{Select ARMA-GARCH model (for marginal models)}
\usage{
select_garch(x)
}
\arguments{
\item{x}{time series.}
}
\value{
A model returned from [fGarch::garchFit()].
}
\description{
Select ARMA-GARCH model (for marginal models)
}
\examples{
fit <- select_garch(rnorm(100))

}
